+++ changelog.md
@@ -1,4 +1,4 @@
-##### Unreleased
+##### Version 1.18.2-0.3
 
 * Plowing machine
   * Plow dirt much quicker using a plow and horse